refactor: define provider adapter layout
