+++ src/application/optional.ts
@@ -1,10 +1,1 @@
-export function optionalProp<Key extends string, Value>(
-  key: Key,
-  value: Value | undefined,
-): { readonly [Property in Key]?: Value } {
-  if (value === undefined) {
-    return {}
-  }
-
-  return { [key]: value } as { readonly [Property in Key]?: Value }
-}
+export { optionalProp } from '../utils/optional.js'
