+++ src/providers/messenger/constants.ts
@@ -1,0 +1,2 @@
+export const TELEGRAM_MINI_APP_PROVIDER_ID = 'telegram-mini-app'
+export const MAX_WEBAPP_PROVIDER_ID = 'max-webapp'
